fix and update map

+++ db/migrate/20161104190236_geolocations.rb
@@ -2,8 +2,8 @@ class Geolocations < ActiveRecord::Migration
   def change
     create_table :geolocations do |t|
       t.string :address
-      t.decimal :lat
-      t.decimal :lng
+      t.integer :lat
+      t.integer :lng
       t.references :user, index: true, foreign_key: true
     end
   end
